+++ qbittorrent/rootfs/etc/cont-init.d/91-qbittorrent_configuration.sh
@@ -6,14 +6,14 @@
 ##########
 
 # Define preferences line
-mkdir -p /config/qBittorrent
+mkdir -p /config/addons_config/qBittorrent
 
 # copy default config
-if [ ! -f /config/qBittorrent/qBittorrent.conf ]; then
-    cp /defaults/qBittorrent.conf /config/qBittorrent/qBittorrent.conf
+if [ ! -f /config/addons_config/qBittorrent/qBittorrent.conf ]; then
+    cp /defaults/qBittorrent.conf /config/addons_config/qBittorrent/qBittorrent.conf
 fi
 
-cd /config/qBittorrent/ || true
+cd /config/addons_config/qBittorrent/ || true
 LINE=$(sed -n '/Preferences/=' qBittorrent.conf)
 LINE=$((LINE + 1))
 
@@ -39,11 +39,12 @@ if bashio::config.has_value 'SavePath'; then
 
     # Replace save path
     CURRENTSAVEPATH=$(sed -n '/Downloads\\SavePath/p' qBittorrent.conf)
-    sed -i "s|${CURRENTSAVEPATH#*=}|$DOWNLOADS|g" qBittorrent.conf 2>/dev/null || true
+    sed -i "s|${CURRENTSAVEPATH#*=}|$DOWNLOADS|g" qBittorrent.conf || \
+    sed -i "${LINE}a Downloads\\SavePath\\$DOWNLOADS" qBittorrent.conf
 
     # Replace session save path
     CURRENTSAVEPATH=$(sed -n '/Session\\DefaultSavePath/p' qBittorrent.conf)
-    sed -i "s|${CURRENTSAVEPATH#*=}|$DOWNLOADS|g" qBittorrent.conf 2>/dev/null || true
+    sed -i "s|${CURRENTSAVEPATH#*=}|$DOWNLOADS|g" qBittorrent.conf || true
 
     # Info
     bashio::log.info "Downloads can be found in $DOWNLOADS"
@@ -100,7 +101,7 @@ fi
 # WHITELIST    #
 ################
 
-cd /config/qBittorrent/ || true
+cd /config/addons_config/qBittorrent/ || true
 if bashio::config.has_value 'whitelist'; then
     WHITELIST=$(bashio::config 'whitelist')
     #clean data
@@ -114,7 +115,7 @@ fi
 # USERNAME    #
 ###############
 
-cd /config/qBittorrent/ || true
+cd /config/addons_config/qBittorrent/ || true
 if bashio::config.has_value 'Username'; then
     USERNAME=$(bashio::config 'Username')
     #clean data
@@ -162,8 +163,8 @@ if bashio::config.has_value 'customUI' && [ ! "$CUSTOMUI" = default ]; then
     rm /webui/*.zip
     CUSTOMUIDIR="$(dirname "$(find /webui/"$CUSTOMUI" -iname "public" -type d)")"
     # Set qbittorrent
-    sed -i "$LINE i\WebUI\\\AlternativeUIEnabled=true" /config/qBittorrent/qBittorrent.conf
-    sed -i "$LINE i\WebUI\\\RootFolder=$CUSTOMUIDIR" /config/qBittorrent/qBittorrent.conf
+    sed -i "$LINE i\WebUI\\\AlternativeUIEnabled=true" /config/addons_config/qBittorrent/qBittorrent.conf
+    sed -i "$LINE i\WebUI\\\RootFolder=$CUSTOMUIDIR" /config/addons_config/qBittorrent/qBittorrent.conf
     # Set nginx
     #sed -i "s=/vuetorrent/public/=$CUSTOMUIDIR/public/=g" /etc/nginx/servers/ingress.conf
     #sed -i "s=vue.torrent=$CUSTOMUI.torrent=g" /etc/nginx/servers/ingress.conf
@@ -175,4 +176,4 @@ fi
 ##########
 
 bashio::log.info "Default username/password : admin/adminadmin"
-bashio::log.info "Configuration can be found in /config/qBittorrent"
+bashio::log.info "Configuration can be found in /config/addons_config/qBittorrent"